Working mongoose implementation
Connects to mongoDB atlas instance as specified in the server.js connectionString.
Please contact Jonathan for access to the instance.

Currently handles get requests, will add more crud in the future.

Contains commented out optional code to create model object and save it to the db, then fetch all entries, as an example